Massive work to get subscribers online.

+++ source/LvStreamStore/EventStreamReader.cs
@@ -6,7 +6,7 @@ public abstract class EventStreamReader : IAsyncEnumerable<StreamItem> {
     public abstract IAsyncEnumerator<StreamItem> GetAsyncEnumerator(CancellationToken cancellationToken = default);
 
     public interface IEnumerator : IAsyncEnumerator<StreamItem> {
-        int Position { get; }
+        long Position { get; }
         int Offset { get; }
     }
 }
